Gestion des scènes à part dans une autre branche

+++ Game.sh
@@ -11,21 +11,29 @@ check_cmd() {
     if [[ $? -eq 0 ]]; then
         if [ -z "$1" ]; then 
             echo -e "${GREEN}OK.${NC}"
-            rm -f "$2" > /dev/null 2>&1
+            if [[ -f "$2" ]]; then
+                rm -f "$2" > /dev/null 2>&1
+            fi
         else 
             echo -e "${GREEN}OK pour $1.${NC}"
-            rm -f "$2" > /dev/null 2>&1
+            if [[ -f "$2" ]]; then
+                rm -f "$2" > /dev/null 2>&1
+            fi
         fi
     else
         if [ -z "$1" ]; then 
             echo -e "${RED}ERREUR !${NC}"
-            cat "$2"
-            rm -f "$2" > /dev/null 2>&1
+            if [[ -f "$2" ]]; then
+                cat "$2"
+                rm -f "$2" > /dev/null 2>&1
+            fi
             exit 1
         else
             echo -e "${RED}ERREUR pour $1.${NC}"
-            cat "$2"
-            rm -f "$2" > /dev/null 2>&1
+            if [[ -f "$2" ]]; then
+                cat "$2"
+                rm -f "$2" > /dev/null 2>&1
+            fi
             exit 1
         fi
     fi
